requirements.txt
streamlit>=1.28.0
requests>=2.31.0
python-dotenv>=1.0.0
geocoder>=1.38.1
google-generativeai>=0.3.0
folium>=0.14.0
streamlit-folium>=0.13.0
pandas>=2.0.0
plotly>=5.15.0
reportlab>=4.0.0